refactor: :recycle: change type of var

+++ internal/domain/rgb.go
@@ -1,5 +1,5 @@
 package domain
 
 type RGB struct {
-	R, G, B uint8
+	R, G, B int
 }
